Implement auto sending position function

+++ app/src/main/java/com/example/maraudersmap/LocationXTO.kt
@@ -1,6 +1,6 @@
 package com.example.maraudersmap
 
 data class LocationXTO (
-    var latitude: Long? = null,
-    var longitude: Long? = null
+    var latitude: Double? = null,
+    var longitude: Double? = null
 )
